Later version of mu vs mass code

+++ scripts/changeAxis.py
@@ -6,7 +6,7 @@ parser = OptionParser(usage="usage: %prog [options]",
                       description="Script to scale signal in inputs files by the cross section for SM Higgs boson. This will transform expectations normalized to 1/pb to yield predictions. The signal samples have to be given as options. ARGS corresponds to the given masses, for which the scaling should be performed.")
 parser.add_option("-i"  ,"--input", dest="input", default="test.root", type="string", help="Input file where to find the signal histograms (or workspaces). [Default: test.root]")
 parser.add_option("-k"  ,"--key", dest="key", default="CMS", type="string",         help="Key to change range")
-parser.add_option("-m"  ,"--min", dest="min", default="50" , type="int",help="Min [Default:50]")
+parser.add_option("-m"  ,"--min", dest="min", default="50" , type="float",help="Min [Default:50]")
 parser.add_option("-v"  ,"--verbose", dest="verbose", default=False, action="store_true", help="increase verbosity. [Default: False]")
 
 # check number of arguments; in case print usage
